+++ src/lib/mina/account.ts
@@ -45,7 +45,10 @@ type FetchedAccount = {
     receive: AuthRequired;
     setDelegate: AuthRequired;
     setPermissions: AuthRequired;
-    setVerificationKey: AuthRequired;
+    setVerificationKey: {
+      auth: AuthRequired;
+      txnVersion: string;
+    };
     setZkappUri: AuthRequired;
     editActionState: AuthRequired;
     setTokenSymbol: AuthRequired;
@@ -83,7 +86,10 @@ const accountQuery = (publicKey: string, tokenId: string) => `{
       receive
       setDelegate
       setPermissions
-      setVerificationKey
+      setVerificationKey {
+        auth
+        txnVersion
+      }
       setZkappUri
       editActionState
       setTokenSymbol